chatbot-backend/requirements.txt
absl-py
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asttokens
attrs
Authlib
backoff
beautifulsoup4
blinker
cachetools
certifi
cffi
chardet
charset-normalizer
chromedriver-autoinstaller
click
colorama
comm
contourpy
cryptography
cssselect
cycler
dataclasses-json
debugpy
decorator
Deprecated
dirtyjson
distro
dnspython
emoji
eval_type_backport
executing
fastapi
fastapi_clerk_auth
feedfinder2
feedparser
filelock
filetype
fonttools
frozenlist
fsspec
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-health-checking
grpcio-status
grpcio-tools
h11
html2text
html5lib
httpcore
httplib2
httpx
huggingface
huggingface-hub
idna
InstructorEmbedding
ipykernel
ipython
jedi
jieba3k  
Jinja2  
jiter  
joblib  
jsonpath-python  
jsonschema  
jsonschema-specifications  
jupyter_client  
jupyter_core  
kiwisolver  
langdetect  
llama-cloud  
llama-index  
llama-index-agent-openai  
llama-index-agent-openai-legacy  
llama-index-cli  
llama-index-core  
llama-index-embeddings-huggingface  
llama-index-embeddings-instructor  
llama-index-embeddings-openai  
llama-index-indices-managed-llama-cloud  
llama-index-llms-gemini  
llama-index-llms-openai  
llama-index-multi-modal-llms-openai  
llama-index-program-openai  
llama-index-question-gen-openai  
llama-index-readers-file  
llama-index-readers-llama-parse  
llama-index-readers-web  
llama-index-vector-stores-weaviate  
llama-parse  
lxml  
Markdown  
markdown-it-py  
MarkupSafe  
marshmallow  
matplotlib  
matplotlib-inline  
mdurl  
motor  
mpmath  
multidict  
mypy-extensions  
narwhals  
ndjson  
nest-asyncio  
networkx  
newspaper3k  
nltk  
numpy  
odmantic  
olefile  
openai  
outcome  
packaging  
pandas  
parso  
pillow  
pip  
platformdirs  
playwright  
prompt_toolkit  
propcache  
proto-plus  
protobuf  
psutil  
pure_eval  
pyarrow  
pyasn1  
pyasn1_modules  
pycparser  
pydantic  
pydantic_core  
pydeck  
pyee  
Pygments  
PyJWT  
pymongo  
pyparsing  
pypdf  
PySocks  
python-dateutil  
python-dotenv  
python-iso639  
python-magic  
python-multipart  
python-oxmsg  
pytz  
pywin32  
PyYAML  
pyzmq  
RapidFuzz  
referencing  
regex  
requests  
requests-file  
requests-toolbelt  
rich  
rpds-py  
rsa  
safetensors  
scikit-learn  
scipy  
selenium  
sentence-transformers  
setuptools  
sgmllib3k  
six  
smmap  
sniffio  
sortedcontainers  
soupsieve  
spider-client  
SQLAlchemy  
stack-data  
starlette  
streamlit  
striprtf  
sympy  
tenacity  
tensorboard  
tensorboard-data-server  
threadpoolctl  
tiktoken  
tinysegmenter  
tldextract  
tokenizers  
toml  
torch  
tornado  
tqdm  
traitlets  
transformers  
trio  
trio-websocket  
typing_extensions  
typing-inspect  
tzdata  
unicorn  
unstructured  
unstructured-client  
uritemplate  
urllib3  
uvicorn  
validators  
watchdog  
wcwidth  
weaviate-client  
webencodings  
websocket-client  
Werkzeug  
wrapt  
wsproto  
yarl  
